mevguard service/unified-mempool-system/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-mempool-system"
version = "1.0.0"
description = "World-class mempool monitoring system consolidating 11 different services into a unified, real-time, synchronized monitoring platform"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Scorpius Team", email = "support@scorpius.ai"}
]
maintainers = [
    {name = "Scorpius Team", email = "support@scorpius.ai"}
]
keywords = [
    "mempool", "blockchain", "ethereum", "mev", "defi", "monitoring", 
    "security", "quantum", "threat-intelligence", "risk-scoring"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "asyncio-mqtt>=0.16.1",
    
    # Web3 and blockchain
    "web3>=6.11.3",
    "eth-account>=0.9.0",
    "eth-typing>=3.5.2",
    "eth-utils>=2.3.0",
    
    # Data processing and analysis
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "scikit-learn>=1.3.2",
    "scipy>=1.11.4",
    
    # Database and caching
    "redis>=5.0.1",
    "aioredis>=2.0.1",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    
    # HTTP and networking
    "aiohttp>=3.9.1",
    "httpx>=0.25.2",
    "websockets>=12.0",
    
    # Monitoring and observability
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    
    # Security and cryptography
    "cryptography>=41.0.8",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Configuration and environment
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    
    # System monitoring
    "psutil>=5.9.6",
    "memory-profiler>=0.61.0",
    
    # Time and date handling
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    
    # JSON and serialization
    "orjson>=3.9.10",
    "msgpack>=1.0.7",
    
    # Machine learning (for quantum analysis simulation)
    "tensorflow>=2.15.0",
    "torch>=2.1.1",
    "transformers>=4.36.0",
    
    # Quantum computing simulation
    "qiskit>=0.45.0",
    "cirq>=1.2.0",
    
    # Additional utilities
    "click>=8.1.7",
    "tqdm>=4.66.1",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
    "httpx>=0.25.2",
]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.2",
]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.8",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/scorpius/unified-mempool-system"
Documentation = "https://unified-mempool-system.readthedocs.io"
Repository = "https://github.com/scorpius/unified-mempool-system.git"
Issues = "https://github.com/scorpius/unified-mempool-system/issues"
Changelog = "https://github.com/scorpius/unified-mempool-system/blob/main/CHANGELOG.md"

[project.scripts]
unified-mempool = "unified_mempool.cli:main"
mempool-monitor = "unified_mempool.cli:monitor"
mempool-api = "unified_mempool.cli:api"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]